requirements.txt
tensorflow-cpu
streamlit
pillow
numpy
scikit-learn
jupyter
Flask